[implemented] current weather response data class

+++ app/src/main/java/com/jvmori/openweather/currentWeather/domain/entities/CurrentWeatherEntity.kt
@@ -3,5 +3,5 @@ package com.jvmori.openweather.currentWeather.domain.entities
 data class CurrentWeatherEntity(
     val city: String,
     val iconUrl: String,
-    val temperature: Float
+    val temperature: Double
 )
